+++ src/main/java/vector_editor/model/Shapes/Pencil.java
@@ -10,7 +10,7 @@ public class Pencil extends ShapeObject {  //for now the separated class from Pe
 
     protected ArrayList<Point> points;
 
-    public Pencil(int x, int y, int x2, int y2, Color c)
+    public Pencil(double x, double y, double x2, double y2, Color c)
     {
         super(x, y, x2, y2, c);
         points= new ArrayList<>();
